--- a/db/commitlog/commitlog.hh
+++ b/db/commitlog/commitlog.hh
@@ -87,8 +87,9 @@ public:
     struct config {
         config() = default;
         config(const config&) = default;
-        static config from_db_config(const db::config&, size_t shard_available_memory);
+        static config from_db_config(const db::config&, seastar::scheduling_group sg, size_t shard_available_memory);
 
+        seastar::scheduling_group sched_group;
         sstring commit_log_location;
         sstring metrics_category_name;
         uint64_t commitlog_total_space_in_mb = 0;