--- a/replica/database.hh
+++ b/replica/database.hh
@@ -1254,6 +1254,7 @@ struct database_config {
     seastar::scheduling_group statement_scheduling_group;
     seastar::scheduling_group streaming_scheduling_group;
     seastar::scheduling_group gossip_scheduling_group;
+    seastar::scheduling_group commitlog_scheduling_group;
     size_t available_memory;
     std::optional<sstables::sstable_version_types> sstables_format;
 };