@raphaelsc 2023-05-23
> Nowadays all commitlog code runs in whatever sched group it's kicked from. 

Isn't this by design?

As you want commit log writes to inherit the priority of the class which invoked them. If I am writing to commit log with a low prio class, I want it to have less bandwidth than writes for a high prio class.

What am I missing?