@raphaelsc 2023-05-23
On Tue, May 23, 2023, 16:08 Pavel Emelyanov ***@***.***>
wrote:

> Nowadays all commitlog code runs in whatever sched group it's kicked from.
>
> Isn't this by design?
>
> I think that commitlog is not CPU-bound, but rather IO-bound, so it
> doesn't change much which sched group it runs in.
>
Ah sorry for the noise. I misread it as io group.


As you want commit log writes to inherit the priority of the class which
> invoked them. If I am writing to commit log with a low prio class, I want
> it to have less bandwidth than writes for a high prio class.
>
> This time no -- IO priority class for commitlog shouldn't be inherited and
> *that* *is* by design for sure.
>
> What am I missing?
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/scylladb/scylladb/pull/14005#issuecomment-1559989454>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AAKYA4Y35WUNQLJFDSMTL4DXHUDKJANCNFSM6AAAAAAYMDR3DU>
> .
> You are receiving this because you commented.Message ID:
> ***@***.***>
>